event update action and some refactoring

+++ client/src/redux/actions/eventActions.js
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { GET_EVENTS, GET_EVENT, GET_EVENTS_FAIL, DELETE_EVENT } from '../types'
+import { GET_EVENTS, GET_EVENT, GET_EVENTS_FAIL, UPDATE_EVENT } from '../types'
 import { setAlert } from './alertActions'
 
 // Get all events
@@ -8,7 +8,7 @@ export const getEvents = () => async dispatch => {
     const res = await axios.get('/api/events')
     dispatch({
       type: GET_EVENTS,
-      payload: res.data.filter(event => new Date(event.date) > new Date())
+      payload: res.data
     })
   } catch (err) {
     console.error(err.response.data)
@@ -57,16 +57,37 @@ export const createEvent = (formData, history) => async dispatch => {
   }
 }
 
+// Update an event
+export const updateEvent = (eventId, formData, history) => async dispatch => {
+  const config = {
+    headers: {
+      'Content-Type': 'application/json'
+    }
+  }
+
+  try {
+    const res = await axios.post(`/api/events/${eventId}`, formData, config)
+    dispatch({
+      type: UPDATE_EVENT,
+      payload: res.data
+    })
+    history.push(`/event/${res.data._id}`)
+  } catch (err) {
+    const errors = err.response.data.errors
+    if (errors) {
+      errors.map(error => dispatch(setAlert(error)))
+    }
+  }
+}
+
 // Delete an event
 export const deleteEvent = (id, history) => async dispatch => {
-  try {
-    await axios.delete(`/api/events/${id}`)
-    dispatch({
-      type: DELETE_EVENT,
-      payload: id
-    })
-    history.push(`/dashboard`)
-  } catch (err) {
-    console.log(err.response.data)
+  if (window.confirm('Es-tu vraiment VRAIMENT sûr de vouloir supprimer ton event ?')) {
+    try {
+      await axios.delete(`/api/events/${id}`)
+      history.push(`/dashboard`)
+    } catch (err) {
+      console.log(err.message)
+    }
   }
 }
